+++ include/Projectile.hpp
@@ -19,6 +19,14 @@ protected:
   float xVector;
   //the y of the vector of movement
   float yVector;
+  //scale applied to the xvector
+  float xVectorScale;
+  //scale applied to the yvector
+  float yVectorScale;
+  //the x coordinate of the taget
+  float xTarget;
+  //the y coordinate of the target
+  float yTarget;
 
   //the unique identifier for the projectile when it is spawned
   int projectileID;
@@ -55,6 +63,20 @@ public:
    * @param y: the y component of the vectors
    */
   void setVector(float x, float y);
+
+  /*
+   * set the positions for the target of the projectile
+   * @param xScale: the x scaling for the vector towards the target
+   * @param yScale: the y scaling for the vector towards the target
+   */
+  void setVectorScale(float xScale, float yScale);
+
+  /*
+   * set the positions for the target of the projectile
+   * @param xpos: the x coordiante of the target of the projectile
+   * @param ypos: the y coordinate of the target of the projectile
+   */
+  void setTargetPos(float x, float y);
 
   /*
    * set the identifier for this projectile