+++ src/Obstacle.cpp
@@ -5,6 +5,7 @@ Obstacle::Obstacle(shared_ptr<EventManager> eventManager,shared_ptr<TextLoader> 
   this -> eventManager = eventManager;
   this -> textLoader = textLoader;
   this -> isMelee = false;
+  this -> radiusVisible = false;
 }
 
 /*